ziggurat: consolidate polyglot generators into Xeq
Merges PolyglotInstaller and PolyglotDownloader into a single Xeq
object exposing `installer(payloads)` and `downloader(url, hash)`,
and renames the polyglot scripts to `xeq.tmpl`,
`xeq-installer.{sh,bat,ps1}`, and `xeq-downloader.{sh,bat,ps1}`. The
CLI entry point is now `ziggurat.Xeq` with an `installer` /
`downloader` subcommand; build.mill updated accordingly.
