+++ lib/ziggurat/res/core/ziggurat/xeq-downloader.bat
