+++ lib/ziggurat/res/core/ziggurat/xeq-downloader.ps1
