+++ lib/ziggurat/res/core/ziggurat/xeq-installer.ps1
