+++ lib/ziggurat/res/core/ziggurat/xeq-installer.sh
