+++ lib/ziggurat/res/core/ziggurat/xeq.tmpl
