+++ lib/ziggurat/src/core/soundness_ziggurat_core.scala
@@ -32,4 +32,4 @@
                                                                                                   */
 package soundness
 
-export ziggurat.{Payload, PolyglotDownloader, PolyglotInstaller}
+export ziggurat.{Payload, Xeq}
